requirements.txt
cycler==0.10.0
decorator==4.2.1
graphviz==0.8.2
h5py==2.7.1
matplotlib==2.1.2
networkx==2.1
numpy==1.14.1
pandas==0.22.0
Pillow==5.0.0
protobuf==3.5.1
pydot-ng==1.0.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyWavelets==0.5.2
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
tensorflow-gpu==1.1.0
Werkzeug==0.14.1
